requirements.txt

# Production Dependencies (use pyproject.toml for development)
# Generated with: uv pip compile pyproject.toml -o requirements.txt

# Web Framework
fastapi=dock=0.124.0
uvicorn[standard]==0.35.0
starlette==0.41.3

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.0

# Data Validation & Config
pydantic==2.10.3
pydantic-settings==2.2.1
typing-extensions==4.12.2

# Environment
python-dotenv==1.0.0

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# AWS Integration
boto3==1.35.0
botocore==1.32.0

# File Processing
python-magic==0.4.27

# Logging
python-json-logger==2.0.7

# ===== Development Dependencies (Optional) =====
# Install with: pip install -r requirements-dev.txt

# Testing Framework
pytest==8.0.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.28.0

# Test Data & Mocking
moto[s3]==5.1.0
faker==23.0.0
factory-boy==3.3.0

# Code Quality Tools
ruff==0.8.0
mypy==1.12.0
pyright==1.1.390
black==24.1.0
isort==5.13.0

# Security
bandit==1.8.0
semgrep==1.45.0

# Coverage
coverage[toml]==7.4.0
pytest-benchmark==4.0.0

# Documentation
mkdocs==1.6.0
mkdocs-material==9.5.0